refactor(auth): move auth logging logic to keychain package
